requirements.txt
# Core dependencies
whoosh>=2.7.4
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.21.0
scipy>=1.7.3

# Text processing
nltk>=3.8
spacy>=3.4.0

# Vector databases (optional alternatives)
chromadb>=0.4.0
pymilvus>=2.3.0

# Database
# sqlite3 is built-in Python

# LLM
groq>=0.4.0
openai>=1.0.0
transformers>=4.21.0

# Utilities
pandas>=1.5.0
tqdm>=4.64.0
scikit-learn>=1.1.0
rank-bm25>=0.2.2

# Configuration
pydantic>=2.0.0
python-dotenv>=1.0.0

# Logging and monitoring
loguru>=0.7.0